backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==19.3.0
black==21.6b0
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
debugpy==1.5.1
defusedxml==0.7.1
Django==3.2.11
django-colorfield==0.4.2
django-cors-headers==3.10.0
django-environ==0.4.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.1.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
more-itertools==8.2.0
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==20.3
pathspec==0.8.1
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.8.6
ptvsd==4.3.2
py==1.8.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2019.3
regex==2021.7.6
reportlab==3.6.2
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
